listen() should take params by reference and return a bool
This updates `listen()` to take params by reference so, as shown in
the `lsp_server` example, the `clone()` can be deferred until `listen()`
verifies it is going to process the notification based on its `method`.

This also changes `listen()` to return a `bool` to indicate whether the
notification was handled by `listen()`. As shown by the updated example, this
makes it more straightforward to efficiently add logic for other notification
types.